bunch of small changes
Made a few changes to the debug command (@debug) intended to make it easier to use. Added the ability to set an item's location to the item editor. Added a function to reload help topics. Kludged a fix for prototypes from a game module not being added to the game's prototypes. Wrote help file for setskill command (@setskill).