bug fixes and other small changes
FIXED problem with setting a status of AFK and getting locked out of the regular command parser. FIXED mail loading issue that resulted in not loading a player's mail file and then overwriting it on saving with whatever was currently in the mailbox. FIXED the SayCommand being broken (not broadcasting to other players in the room) by changes over time in the way Message objects are handled. RESOLVED non-working tracking of player idle time. UPDATED a couple help files/